requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
numpy==2.3.2
PyJWT==2.10.1
sqlparse==0.5.3
